Updating the branch

+++ src/test/java/ru/skypro/homework/mapper/AdMapperTest.java
@@ -17,7 +17,7 @@ public class AdMapperTest {
     @Test
     public void testAdToAdDTO() {
         User user = new User();
-        user.setId(1);
+        user.setId(1L);
         user.setFirstName("Мария");
         user.setLastName("Синявская");
 
@@ -32,7 +32,7 @@ public class AdMapperTest {
         AdDTO adDTO = adMapper.adToAdDTO(ad);
 
         assertEquals(ad.getPk().intValue(), adDTO.getPk());
-        assertEquals(ad.getUser().getId(), adDTO.getAuthor());
+        assertEquals(ad.getUser().getId(), adDTO.getAuthor().longValue());
         assertEquals("/image/" + ad.getImage().getId(), adDTO.getImage());
         }
